Module aws_sdk_sqs::error

Modules

batch_entry_ids_not_distinct

See BatchEntryIdsNotDistinct

batch_request_too_long

See BatchRequestTooLong

empty_batch_request

See EmptyBatchRequest

invalid_attribute_name

See InvalidAttributeName

invalid_batch_entry_id

See InvalidBatchEntryId

invalid_id_format

See InvalidIdFormat

invalid_message_contents

See InvalidMessageContents

message_not_inflight

See MessageNotInflight

over_limit

See OverLimit

purge_queue_in_progress

See PurgeQueueInProgress

queue_deleted_recently

See QueueDeletedRecently

queue_does_not_exist

See QueueDoesNotExist

queue_name_exists

See QueueNameExists

receipt_handle_is_invalid

See ReceiptHandleIsInvalid

too_many_entries_in_batch_request

See TooManyEntriesInBatchRequest

unsupported_operation

See UnsupportedOperation

Structs

AddPermissionError

BatchEntryIdsNotDistinct

Two or more batch entries in the request have the same Id.

BatchRequestTooLong

The length of all the messages put together is more than the limit.

ChangeMessageVisibilityBatchError

ChangeMessageVisibilityError

CreateQueueError

DeleteMessageBatchError

DeleteMessageError

DeleteQueueError

EmptyBatchRequest

The batch request doesn't contain any entries.

GetQueueAttributesError

GetQueueUrlError

InvalidAttributeName

The specified attribute doesn't exist.

InvalidBatchEntryId

The Id of a batch entry in a batch request doesn't abide by the specification.

InvalidIdFormat

The specified receipt handle isn't valid for the current version.

InvalidMessageContents

The message contains characters outside the allowed set.

ListDeadLetterSourceQueuesError

ListQueueTagsError

ListQueuesError

MessageNotInflight

The specified message isn't in flight.

OverLimit

The specified action violates a limit. For example, ReceiveMessage

PurgeQueueError

PurgeQueueInProgress

Indicates that the specified queue previously received a PurgeQueue request within the last 60 seconds (the time it can take to delete the messages in the queue).

QueueDeletedRecently

You must wait 60 seconds after deleting a queue before you can create another queue

QueueDoesNotExist

The specified queue doesn't exist.

QueueNameExists

A queue with this name already exists. Amazon SQS returns this error only if the request

ReceiptHandleIsInvalid

The specified receipt handle isn't valid.

ReceiveMessageError

RemovePermissionError

SendMessageBatchError

SendMessageError

SetQueueAttributesError

TagQueueError

TooManyEntriesInBatchRequest

The batch request contains more entries than permissible.

UnsupportedOperation

Error code 400. Unsupported operation.

UntagQueueError

Enums

AddPermissionErrorKind

ChangeMessageVisibilityBatchErrorKind

ChangeMessageVisibilityErrorKind

CreateQueueErrorKind

DeleteMessageBatchErrorKind

DeleteMessageErrorKind

DeleteQueueErrorKind

GetQueueAttributesErrorKind

GetQueueUrlErrorKind

ListDeadLetterSourceQueuesErrorKind

ListQueueTagsErrorKind

ListQueuesErrorKind

PurgeQueueErrorKind

ReceiveMessageErrorKind

RemovePermissionErrorKind

SendMessageBatchErrorKind

SendMessageErrorKind

SetQueueAttributesErrorKind

TagQueueErrorKind

UntagQueueErrorKind