Struct aws_sdk_sqs::input::receive_message_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ReceiveMessageInput

Implementations

impl Builder

pub fn queue_url(self, input: impl Into<String>) -> Self

The URL of the Amazon SQS queue from which messages are received.

Queue URLs and names are case-sensitive.

pub fn set_queue_url(self, input: Option<String>) -> Self

pub fn attribute_names(self, input: impl Into<QueueAttributeName>) -> Self

pub fn set_attribute_names(self, input: Option<Vec<QueueAttributeName>>) -> Self

pub fn message_attribute_names(self, input: impl Into<String>) -> Self

pub fn set_message_attribute_names(self, input: Option<Vec<String>>) -> Self

pub fn max_number_of_messages(self, input: i32) -> Self

The maximum number of messages to return. Amazon SQS never returns more messages than this value (however, fewer messages might be returned). Valid values: 1 to 10. Default: 1.

pub fn set_max_number_of_messages(self, input: Option<i32>) -> Self

pub fn visibility_timeout(self, input: i32) -> Self

The duration (in seconds) that the received messages are hidden from subsequent retrieve requests after being retrieved by a ReceiveMessage request.

pub fn set_visibility_timeout(self, input: Option<i32>) -> Self

pub fn wait_time_seconds(self, input: i32) -> Self

The duration (in seconds) for which the call waits for a message to arrive in the queue before returning. If a message is available, the call returns sooner than WaitTimeSeconds. If no messages are available and the wait time expires, the call returns successfully with an empty list of messages.

To avoid HTTP errors, ensure that the HTTP response timeout for ReceiveMessage requests is longer than the WaitTimeSeconds parameter. For example, with the Java SDK, you can set HTTP transport settings using the NettyNioAsyncHttpClient for asynchronous clients, or the ApacheHttpClient for synchronous clients.

pub fn set_wait_time_seconds(self, input: Option<i32>) -> Self

pub fn receive_request_attempt_id(self, input: impl Into<String>) -> Self

This parameter applies only to FIFO (first-in-first-out) queues.

The token used for deduplication of ReceiveMessage calls. If a networking issue occurs after a ReceiveMessage action, and instead of a response you receive a generic error, it is possible to retry the same action with an identical ReceiveRequestAttemptId to retrieve the same set of messages, even if their visibility timeout has not yet expired.

You can use ReceiveRequestAttemptId only for 5 minutes after a ReceiveMessage action.

When you set FifoQueue, a caller of the ReceiveMessage action can provide a ReceiveRequestAttemptId explicitly.

If a caller of the ReceiveMessage action doesn't provide a ReceiveRequestAttemptId, Amazon SQS generates a ReceiveRequestAttemptId.

It is possible to retry the ReceiveMessage action with the same ReceiveRequestAttemptId if none of the messages have been modified (deleted or had their visibility changes).

During a visibility timeout, subsequent calls with the same ReceiveRequestAttemptId return the same messages and receipt handles. If a retry occurs within the deduplication interval, it resets the visibility timeout. For more information, see Visibility Timeout in the Amazon SQS Developer Guide.

If a caller of the ReceiveMessage action still processes messages when the visibility timeout expires and messages become visible, another worker consuming from the same queue can receive the same messages and therefore process duplicates. Also, if a consumer whose message processing time is longer than the visibility timeout tries to delete the processed messages, the action fails with an error.

To mitigate this effect, ensure that your application observes a safe threshold before the visibility timeout expires and extend the visibility timeout as necessary.

While messages with a particular MessageGroupId are invisible, no more messages belonging to the same MessageGroupId are returned until the visibility timeout expires. You can still receive messages with another MessageGroupId as long as it is also visible.

If a caller of ReceiveMessage can't track the ReceiveRequestAttemptId, no retries work until the original visibility timeout expires. As a result, delays might occur but the messages in the queue remain in a strict order.

The maximum length of ReceiveRequestAttemptId is 128 characters. ReceiveRequestAttemptId can contain alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

For best practices of using ReceiveRequestAttemptId, see Using the ReceiveRequestAttemptId Request Parameter in the Amazon SQS Developer Guide.

pub fn set_receive_request_attempt_id(self, input: Option<String>) -> Self

pub fn build(self) -> Result<ReceiveMessageInput, BuildError>

Consumes the builder and constructs a ReceiveMessageInput