Module aws_sdk_sqs::model

Modules

batch_result_error_entry

See BatchResultErrorEntry

change_message_visibility_batch_request_entry

See ChangeMessageVisibilityBatchRequestEntry

change_message_visibility_batch_result_entry

See ChangeMessageVisibilityBatchResultEntry

delete_message_batch_request_entry

See DeleteMessageBatchRequestEntry

delete_message_batch_result_entry

See DeleteMessageBatchResultEntry

message

See Message

message_attribute_value

See MessageAttributeValue

message_system_attribute_value

See MessageSystemAttributeValue

send_message_batch_request_entry

See SendMessageBatchRequestEntry

send_message_batch_result_entry

See SendMessageBatchResultEntry

Structs

BatchResultErrorEntry

Gives a detailed description of the result of an action on each entry in the

ChangeMessageVisibilityBatchRequestEntry

Encloses a receipt handle and an entry id for each message in

ChangeMessageVisibilityBatchResultEntry

Encloses the Id of an entry in

DeleteMessageBatchRequestEntry

Encloses a receipt handle and an identifier for it.

DeleteMessageBatchResultEntry

Encloses the Id of an entry in

Message

An Amazon SQS message.

MessageAttributeValue

The user-specified message attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see

MessageSystemAttributeValue

The user-specified message system attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see

SendMessageBatchRequestEntry

Contains the details of a single Amazon SQS message along with an Id.

SendMessageBatchResultEntry

Encloses a MessageId for a successfully-enqueued message in a

Enums

MessageSystemAttributeName

MessageSystemAttributeNameForSends

QueueAttributeName