Module aws_sdk_sqs::operation

Structs

AddPermission

Adds a permission to a queue for a specific

ChangeMessageVisibility

Changes the visibility timeout of a specified message in a queue to a new value. The

ChangeMessageVisibilityBatch

Changes the visibility timeout of multiple messages. This is a batch version of

CreateQueue

Creates a new standard or FIFO queue. You can pass one or more attributes in

DeleteMessage

Deletes the specified message from the specified queue. To select the message to

DeleteMessageBatch

Deletes up to ten messages from the specified queue. This is a batch version of

DeleteQueue

Deletes the queue specified by the QueueUrl, regardless of the queue's contents.

GetQueueAttributes

Gets attributes for the specified queue.

GetQueueUrl

Returns the URL of an existing Amazon SQS queue.

ListDeadLetterSourceQueues

Returns a list of your queues that have the RedrivePolicy queue attribute configured with a dead-letter queue.

ListQueueTags

List all cost allocation tags added to the specified Amazon SQS queue. For an overview, see Tagging

ListQueues

Returns a list of your queues in the current region. The response includes a maximum of 1,000 results. If you specify a value for the optional

PurgeQueue

Deletes the messages in a queue specified by the QueueURL

ReceiveMessage

Retrieves one or more messages (up to 10), from the specified queue. Using the WaitTimeSeconds parameter enables long-poll support.

RemovePermission

Revokes any permissions in the queue policy that matches the specified Label parameter.

SendMessage

Delivers a message to the specified queue.

SendMessageBatch

Delivers up to ten messages to the specified queue. This is a batch version of

SetQueueAttributes

Sets the value of one or more queue attributes. When you change a queue's attributes, the change can take up to 60 seconds for most of the attributes to propagate throughout the Amazon SQS system.

TagQueue

Add cost allocation tags to the specified Amazon SQS queue. For an overview, see Tagging

UntagQueue

Remove cost allocation tags from the specified Amazon SQS queue. For an overview, see Tagging