Module aws_sdk_sqs::output

Modules

add_permission_output

See AddPermissionOutput

change_message_visibility_batch_output

See ChangeMessageVisibilityBatchOutput

change_message_visibility_output

See ChangeMessageVisibilityOutput

create_queue_output

See CreateQueueOutput

delete_message_batch_output

See DeleteMessageBatchOutput

delete_message_output

See DeleteMessageOutput

delete_queue_output

See DeleteQueueOutput

get_queue_attributes_output

See GetQueueAttributesOutput

get_queue_url_output

See GetQueueUrlOutput

list_dead_letter_source_queues_output

See ListDeadLetterSourceQueuesOutput

list_queue_tags_output

See ListQueueTagsOutput

list_queues_output

See ListQueuesOutput

purge_queue_output

See PurgeQueueOutput

receive_message_output

See ReceiveMessageOutput

remove_permission_output

See RemovePermissionOutput

send_message_batch_output

See SendMessageBatchOutput

send_message_output

See SendMessageOutput

set_queue_attributes_output

See SetQueueAttributesOutput

tag_queue_output

See TagQueueOutput

untag_queue_output

See UntagQueueOutput

Structs

AddPermissionOutput

ChangeMessageVisibilityBatchOutput

For each message in the batch, the response contains a

ChangeMessageVisibilityOutput

CreateQueueOutput

Returns the QueueUrl attribute of the created queue.

DeleteMessageBatchOutput

For each message in the batch, the response contains a

DeleteMessageOutput

DeleteQueueOutput

GetQueueAttributesOutput

A list of returned queue attributes.

GetQueueUrlOutput

For more information, see Interpreting Responses in the Amazon SQS Developer Guide.

ListDeadLetterSourceQueuesOutput

A list of your dead letter source queues.

ListQueueTagsOutput

ListQueuesOutput

A list of your queues.

PurgeQueueOutput

ReceiveMessageOutput

A list of received messages.

RemovePermissionOutput

SendMessageBatchOutput

For each message in the batch, the response contains a

SendMessageOutput

The MD5OfMessageBody and MessageId elements.

SetQueueAttributesOutput

TagQueueOutput

UntagQueueOutput